FAC-398:feature:Rename to AssessmentUserRole
