--- flickit-assessment-data/src/main/java/org/flickit/assessment/data/jpa/users/userassessmentrole/UserAssessmentRoleJpaRepository.java
@@ -1,7 +1,0 @@
-package org.flickit.assessment.data.jpa.users.userassessmentrole;
-
-import org.flickit.assessment.data.jpa.users.userassessmentrole.UserAssessmentRoleJpaEntity.EntityId;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserAssessmentRoleJpaRepository extends JpaRepository<UserAssessmentRoleJpaEntity, EntityId> {
-}
