+++ flickit-assessment-data/src/main/resources/liquibase/changelog-0.19.xml
@@ -5,7 +5,7 @@
                                        http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.17.xsd">
 
     <changeSet id="0.19-01" author="Mahdi Rezazadeh">
-        <createTable tableName="fau_user_assessment_role">
+        <createTable tableName="fac_assessment_user_role">
             <column name="assessment_id" type="bigint">
                 <constraints nullable="false"/>
             </column>
@@ -19,16 +19,16 @@
     </changeSet>
 
     <changeSet id="0.19-02" author="Mahdi Rezazadeh">
-        <addPrimaryKey tableName="fau_user_assessment_role"
+        <addPrimaryKey tableName="fac_assessment_user_role"
                        columnNames="assessment_id, user_id"/>
     </changeSet>
 
     <changeSet id="0.19-03" author="Mahdi Rezazadeh">
-        <addForeignKeyConstraint baseTableName="fau_user_assessment_role"
+        <addForeignKeyConstraint baseTableName="fac_assessment_user_role"
                                  baseColumnNames="user_id"
                                  referencedTableName="account_user"
                                  referencedColumnNames="id"
-                                 constraintName="fk_fau_user_assessment_role_account_user"
+                                 constraintName="fk_fac_assessment_user_role_account_user"
                                  deferrable="true" initiallyDeferred="true" onDelete="CASCADE"
                                  validate="true"/>
     </changeSet>
